+++ src/rescuetime2gcal/commands/cli.py
@@ -1,0 +1,3 @@
+import typer
+
+app = typer.Typer()
