+++ src/rescuetime2gcal/sources/rescuetime.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Sequence
 import devtools
 import pytz
 import requests
-from rescuetime2gcal.source_event import BareEvent
+from rescuetime2gcal.sources.source_event import BareEvent
 
 if TYPE_CHECKING:
     import pytz.tzfile
