+++ src/rescuetime2gcal/sources/source.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING, Protocol, Sequence
 
 if TYPE_CHECKING:
-    from rescuetime2gcal.source_event import SourceEvent
+    from rescuetime2gcal.sources.source_event import SourceEvent
 
 
 class EventSource(Protocol):
