--- src/rescuetime2gcal/test_main.py
@@ -1,26 +1,0 @@
-import datetime
-from typing import TYPE_CHECKING, Sequence
-
-from rescuetime2gcal import delta
-from rescuetime2gcal.__main__ import _pipeline
-from rescuetime2gcal.test_preprocessing import FakeDestinationEvent
-
-if TYPE_CHECKING:
-    from rescuetime2gcal.preprocessing import DestinationEvent
-
-
-def test_pipeline_should_remove_duplicates():
-    def destination_client() -> Sequence["DestinationEvent"]:
-        return [FakeDestinationEvent(id="0"), FakeDestinationEvent(id="1")]
-
-    changes = _pipeline(
-        source_events=[],
-        destination_events=destination_client(),
-        exclude_titles=[],
-        metadata_enrichment=[],
-        category2emoji={},
-        min_duration=datetime.timedelta(days=1),
-        merge_gap=datetime.timedelta(days=1),
-        exclude_apps=[],
-    )
-    assert changes == [delta.DeleteEvent(event=FakeDestinationEvent(id="1"))]
